+++ src/main/java/fr/eni/tp/filmotheque/bll/jpa/ParticipantServiceImpl.java
@@ -1,10 +1,14 @@
 package fr.eni.tp.filmotheque.bll.jpa;
 
 import fr.eni.tp.filmotheque.bll.ParticipantService;
+import fr.eni.tp.filmotheque.bo.Genre;
 import fr.eni.tp.filmotheque.bo.Participant;
 import fr.eni.tp.filmotheque.dal.ParticipantRepository;
+import jakarta.servlet.http.Part;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -31,6 +35,19 @@ public class ParticipantServiceImpl implements ParticipantService {
 
     @Override
     public void supprimerParticipantParId(Long id) {
-        participantRepository.deleteById(id);
+        if (participantRepository.existsById(id)){
+            participantRepository.deleteById(id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Le participant n'a pas été trouvé");
+        }
+    }
+
+    @Override
+    public void putParticipant(Participant participant) {
+        if (participantRepository.existsById(participant.getId())){
+            participantRepository.save(participant);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Le participant n'a pas été trouvé");
+        }
     }
 }
